(PC-20344) HomeHeader: extract activation banner
